@@ -1,7 +1,7 @@
 ﻿<Project Sdk="Microsoft.NET.Sdk">
 
     <PropertyGroup>
-        <TargetFramework>netcoreapp3.1</TargetFramework>
+        <TargetFramework>net5.0</TargetFramework>
         <TreatWarningsAsErrors>true</TreatWarningsAsErrors>
         <Title>VxSort</Title>
         <Authors>damageboy</Authors>
@@ -22,25 +22,28 @@
         <PackageIconUrl>..\vxsort.png</PackageIconUrl>
     </PropertyGroup>
     <ItemGroup>
-      <PackageReference Include="Fody" />
-      <PackageReference Include="LocalsInit.Fody" />
       <PackageReference Include="Microsoft.CodeAnalysis.FxCopAnalyzers" />
       <PackageReference Include="Microsoft.CodeQuality.Analyzers" />
       <PackageReference Include="Microsoft.SourceLink.GitHub" />
     </ItemGroup>
     <ItemGroup>
       <None Remove="vxsort.svg" />
-      <None Update="BitonicSort.Generated.tt">
+      <None Update="Reference\BitonicSort.Generated.tt">
         <Generator>TextTemplatingFileGenerator</Generator>
         <LastGenOutput>BitonicSort.Generated.cs</LastGenOutput>
       </None>
     </ItemGroup>
     <ItemGroup>
-      <Compile Update="BitonicSort.Generated.cs">
+      <Compile Update="Reference\BitonicSort.Generated.cs">
         <DependentUpon>BitonicSort.Generated.tt</DependentUpon>
+        <DesignTime>True</DesignTime>
+        <AutoGen>True</AutoGen>
       </Compile>
     </ItemGroup>
     <ItemGroup>
       <Folder Include="codegen\" />
+    </ItemGroup>
+    <ItemGroup>
+      <Service Include="{508349b6-6b84-4df5-91f0-309beebad82d}" />
     </ItemGroup>
 </Project>
