@@ -1,110 +1,40 @@
-using System;
+﻿using System;
+using System.Collections.Generic;
 using System.Runtime.CompilerServices;
 using System.Runtime.Intrinsics;
-using static System.Runtime.Intrinsics.X86.Avx2;
 
 namespace VxSort
 {
-    using V = Vector256<int>;
-    static partial class BitonicSort<T> where T : unmanaged, IComparable<T>
+    static unsafe partial class BitonicSort
     {
+        // * We might read the last 4 bytes into a 128-bit vector for 64-bit element masking
+        // * We might read the last 8 bytes into a 128-bit vector for 32-bit element masking
+        // This mostly applies to debug mode, since without optimizations, most compilers
+        // actually execute the instruction stream _mm256_cvtepi8_epiNN + _mm_loadu_si128 as they are given.
+        // In contract, release/optimizing compilers, turn that very specific instruction pair to
+        // a more reasonable: vpmovsxbq ymm0, dword [rax*4 + mask_table_4], eliminating the 128-bit
+        // load completely and effectively reading 4/8 (depending if the instruction is vpmovsxb[q,d]
 
-        // Legend:
-        // X - shuffle/permute mask for generating a cross (X) shuffle
-        //     the numbers (1,2,4) denote the stride of the shuffle operation
-        // B - Blend mask, used for blending two vectors according to a given order
-        //     the numbers (1,2,4) denote the "stride" of blending, e.g. 1 means switch vectors
-        //     every element, 2 means switch vectors every two elements and so on...
-        // P - Permute mask, read specific comment about it below...
-        const byte X_1 = 0b10_11_00_01;
-        const byte X_2 = 0b01_00_11_10;
-        const byte B_1 = 0b10_10_10_10;
-        const byte B_2 = 0b11_00_11_00;
-        const byte B_4 = 0b11_11_00_00;
+        public static ReadOnlySpan<byte> mask_table_4 => new byte[]{
+            0xFF, 0xFF, 0xFF, 0xFF,  // 0b0000 (0)
+            0xFF, 0x00, 0x00, 0x00,  // 0b0001 (1)
+            0xFF, 0xFF, 0x00, 0x00,  // 0b0011 (3)
+            0xFF, 0xFF, 0xFF, 0x00,  // 0b0111 (7)
+            0xCC, 0xCC, 0xCC, 0xCC,  // Garbage to make ASAN happy
+            0xCC, 0xCC, 0xCC, 0xCC,  // Garbage to make ASAN happy
+            0xCC, 0xCC, 0xCC, 0xCC,  // Garbage to make ASAN happy
+        };
 
-        // Shuffle (X_R) + Permute (P_X) is a more efficient way
-        // (copied shamelessly from LLVM through compiler explorer)
-        // For implementing X_4, which requires a cross 128-bit lane operation.
-        // A Shuffle (1c lat / 1c tp) + 64 bit permute (3c lat / 1c tp) take 1 more cycle to execute than the
-        // the alternative: PermuteVar8x32 / VPERMD which takes (3c lat / 1c tp)
-        // But, the latter requires loading the permutation entry from cache, which can take up to 5 cycles (when cached)
-        // and costs one more register, which steals a register from us for high-count bitonic sorts.
-        // In short, it's faster this way, from my attempts...
-        const byte X_R = 0b00_01_10_11;
-        const byte P_X = 0b01_00_11_10;
-
-        // Basic 8-element bitonic sort
-        // This will get composed and inlined throughout
-        // the various bitonic-sort sizes:
-        // BitonicSort1V will be directly embedded in BitonicSort{2,3,5,9}V
-        // BitonicSort2V will be directly embedded in BitonicSort{3,4,6,10}V
-        // BitonicSort3V will be directly embedded in BitonicSort{7,11}V
-        // etc.
-        [MethodImpl(MethodImplOptions.AggressiveInlining | MethodImplOptions.AggressiveOptimization)]
-        static void BitonicSort01V(ref V d)
-        {
-            // ReSharper disable JoinDeclarationAndInitializer
-            V min, max, s;
-            // ReSharper restore JoinDeclarationAndInitializer
-            s   = Shuffle(d, X_1);
-            min = Min(s, d);
-            max = Max(s, d);
-            d   = Blend(min, max, B_1);
-
-            s   = Shuffle(d, X_R);
-            min = Min(s, d);
-            max = Max(s, d);
-            d   = Blend(min, max,B_2);
-
-            s   = Shuffle(d, X_1);
-            min = Min(s, d);
-            max = Max(s, d);
-            d   = Blend(min, max,B_1);
-
-            s   = Shuffle(d, X_R);
-            s   = Permute4x64(s.AsInt64(),P_X).AsInt32();
-            min = Min(s, d);
-            max = Max(s, d);
-            d   = Blend(min, max, B_4);
-
-            s   = Shuffle(d, X_2);
-            min = Min(s, d);
-            max = Max(s, d);
-            d   = Blend(min, max,B_2);
-
-            s   = Shuffle(d, X_1);
-            min = Min(s, d);
-            max = Max(s, d);
-            d   = Blend(min, max,  B_1);
-        }
-
-        // Basic bitonic 8-element merge
-        // This will get composed and inlined throughout
-        // the code base for merging larger sized bitonic sorts temporary result states
-        // BitonicSort1VFinish used for BitonicSort{2,3,5,9}V{,Finish}
-        // BitonicSort2VFinish used for BitonicSort{3,4,6,10}V{,Finish}
-
-        [MethodImpl(MethodImplOptions.AggressiveInlining | MethodImplOptions.AggressiveOptimization)]
-        static void BitonicSort01VMerge(ref V d)
-        {
-            // ReSharper disable JoinDeclarationAndInitializer
-            V min, max, s;
-            // ReSharper restore JoinDeclarationAndInitializer
-
-            s   = Permute4x64(d.AsInt64(), P_X).AsInt32();
-            min = Min(s, d);
-            max = Max(s, d);
-            d   = Blend(min, max, B_4);
-
-            s   = Shuffle(d, X_2);
-            min = Min(s, d);
-            max = Max(s, d);
-            d   = Blend(min, max, B_2);
-
-            s   = Shuffle(d, X_1);
-            min = Min(s, d);
-            max = Max(s, d);
-            d   = Blend(min, max, B_1);
-        }
+        public static ReadOnlySpan<byte> mask_table_8 => new byte[]{
+            0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, // 0b00000000 (  0)
+            0xFF, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, // 0b00000001 (  1)
+            0xFF, 0xFF, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, // 0b00000011 (  3)
+            0xFF, 0xFF, 0xFF, 0x00, 0x00, 0x00, 0x00, 0x00, // 0b00000111 (  7)
+            0xFF, 0xFF, 0xFF, 0xFF, 0x00, 0x00, 0x00, 0x00, // 0b00001111 ( 15)
+            0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0x00, 0x00, 0x00, // 0b00011111 ( 31)
+            0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0x00, 0x00, // 0b00111111 ( 63)
+            0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0x00, // 0b01111111 (127)
+            0xCC, 0xCC, 0xCC, 0xCC, 0xCC, 0xCC, 0xCC, 0xCC, // Garbage to make ASAN happy
+        };
     }
 }
