@@ -1,5 +1,5 @@
-﻿// <auto-generated>
-//     This code was generated by a tool on 2020-01-20-15:08:20
+// <auto-generated>
+//     This code was generated by a tool on 2021-10-04-17:49:59
 //
 //     Changes to this file may cause incorrect behavior and will be lost if
 //     the code is regenerated.
@@ -12,7 +12,7 @@ using System.Runtime.Intrinsics;
 using static System.Runtime.Intrinsics.X86.Avx;
 using static System.Runtime.Intrinsics.X86.Avx2;
 
-namespace VxSort
+namespace VxSort.Reference
 {
     using V = Vector256<int>;
     static unsafe partial class BitonicSort<T>
